Fix modules separation
