+++ js/main.js
@@ -1,4 +1,4 @@
-import {getSimilarStays} from './get-similar-stays.js';
+import {getSimilarStays} from './data.js';
 
 const SIMILAR_STAYS_COUNT = 10;
 getSimilarStays(SIMILAR_STAYS_COUNT);
